fix: Separate country utils
