+++ src/landscape/components/LandscapeList.js
@@ -27,7 +27,6 @@ import { withProps } from 'react-hoc';
 import ExternalLink from 'common/components/ExternalLink';
 import TableResponsive from 'common/components/TableResponsive';
 import { useDocumentDescription, useDocumentTitle } from 'common/document';
-import { countryNameForCode } from 'common/utils';
 import PageContainer from 'layout/PageContainer';
 import PageHeader from 'layout/PageHeader';
 import PageLoader from 'layout/PageLoader';
@@ -39,6 +38,7 @@ import { fetchLandscapes } from 'landscape/landscapeSlice';
 import LandscapeMemberLeave from 'landscape/membership/components/LandscapeMemberLeave';
 
 import LandscapeListMap from './LandscapeListMap';
+import { countryNameForCode } from 'common/countries';
 
 const MemberLeaveButton = withProps(LandscapeMemberLeave, {
   label: 'landscape.list_leave_button',
